backend/requirements.txt
fastapi>=0.110.0
uvicorn>=0.27.0
chromadb>=0.4.24
langchain>=0.1.12
langchain-openai>=0.0.8
langchain-community>=0.0.28
pypdf>=4.1.0
python-multipart>=0.0.9
python-dotenv>=1.0.1
